split main file, fix package.json
