+++ package.json
@@ -19,11 +19,9 @@
     "sinon": "^1.16.1",
     "stats-lite": "^2.1.1"
   },
-  {
-    "engines": {
-      "npm": "^4.0.0",
-      "node": ">=6.0.0"
-    }
+  "engines": {
+    "npm": "^4.0.0",
+    "node": ">=6.0.0"
   },
   "main": "index",
   "scripts": {
